requirements.txt
pdf2image==1.17.0
pytesseract
Pillow
googletrans==4.0.0-rc1
